+++ src/auth/auth.service.ts
@@ -4,7 +4,6 @@ import {
   HttpException,
   HttpStatus,
   Injectable,
-  NotFoundException,
   UnauthorizedException,
 } from '@nestjs/common';
 import ms from 'ms';
@@ -535,42 +534,5 @@ export class AuthService {
       refreshToken,
       tokenExpires,
     };
-  }
-
-  async subscribe(currentUserId: number, targetUserId: number): Promise<any> {
-    // Check if there is a user to subscribe to.
-    const targetUser = await this.usersService.findOne({ id: targetUserId });
-
-    if (!targetUser) {
-      throw new NotFoundException('User not found');
-    }
-    // Check that the user is not subscribing to himself
-    if (currentUserId === targetUserId) {
-      throw new NotFoundException('You cannot subscribe to yourself!');
-    }
-
-    const currentUser = await this.usersService.findOne({ id: currentUserId });
-
-    if (!currentUser) {
-      throw new NotFoundException('Current user not found');
-    }
-
-    if (!currentUser.subscribers) {
-      currentUser.subscribers = [];
-    }
-
-    const isAlreadySubscribed = currentUser.subscribers.some(
-      (subscriber) => subscriber.id === targetUserId,
-    );
-
-    if (isAlreadySubscribed) {
-      throw new ConflictException('You are already subscribed to this user');
-    }
-
-    currentUser.subscribers.push(targetUser);
-
-    await currentUser.save();
-
-    return currentUser;
   }
 }
