+++ src/users/entities/user.entity.ts
@@ -111,12 +111,15 @@ export class User extends EntityHelper {
   hash: string | null;
 
   @CreateDateColumn()
+  @Exclude({ toPlainOnly: true })
   createdAt: Date;
 
   @UpdateDateColumn()
+  @Exclude({ toPlainOnly: true })
   updatedAt: Date;
 
   @DeleteDateColumn()
+  @Exclude({ toPlainOnly: true })
   deletedAt: Date;
 
   @OneToMany(() => PostEntity, (post) => post.author)
