+++ src/users/users.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { EntityCondition } from 'src/utils/types/entity-condition.type';
 import { IPaginationOptions } from 'src/utils/types/pagination-options';
@@ -47,5 +51,42 @@ export class UsersService {
 
   async softDelete(id: User['id']): Promise<void> {
     await this.usersRepository.softDelete(id);
+  }
+
+  async subscribe(currentUserId: number, targetUserId: number): Promise<any> {
+    // Check if there is a user to subscribe to.
+    const targetUser = await this.findOne({ id: targetUserId });
+
+    if (!targetUser) {
+      throw new NotFoundException('User not found');
+    }
+    // Check that the user is not subscribing to himself
+    if (currentUserId === targetUserId) {
+      throw new NotFoundException('You cannot subscribe to yourself!');
+    }
+
+    const currentUser = await this.findOne({ id: currentUserId });
+
+    if (!currentUser) {
+      throw new NotFoundException('Current user not found');
+    }
+
+    if (!currentUser.subscribers) {
+      currentUser.subscribers = [];
+    }
+
+    const isAlreadySubscribed = currentUser.subscribers.some(
+      (subscriber) => subscriber.id === targetUserId,
+    );
+
+    if (isAlreadySubscribed) {
+      throw new ConflictException('You are already subscribed to this user');
+    }
+
+    currentUser.subscribers.push(targetUser);
+
+    await currentUser.save();
+
+    return currentUser;
   }
 }
